refactor: extract command handlers from cli.py into synapse/commands/
