+++ synapse/cli.py
@@ -16,21 +16,62 @@ import sys
 import time
 from collections.abc import Callable
 from pathlib import Path
-from typing import TYPE_CHECKING, Literal, cast
+from typing import Literal, cast
 
 import yaml
 
-if TYPE_CHECKING:
-    from synapse.history import HistoryManager
-
-from synapse.a2a_client import A2AClient, get_client
 from synapse.agent_profiles import (
     AgentProfileError,
     AgentProfileStore,
     suggest_petname_ids,
 )
 from synapse.auth import generate_api_key
+from synapse.commands import history as history_commands
+from synapse.commands import memory as memory_commands
+from synapse.commands import messaging as messaging_commands
+from synapse.commands.external import (
+    cmd_external_add,
+    cmd_external_info,
+    cmd_external_list,
+    cmd_external_remove,
+    cmd_external_send,
+)
+from synapse.commands.file_safety_cmd import (
+    cmd_file_safety_cleanup,
+    cmd_file_safety_cleanup_locks,
+    cmd_file_safety_debug,
+    cmd_file_safety_history,
+    cmd_file_safety_lock,
+    cmd_file_safety_locks,
+    cmd_file_safety_recent,
+    cmd_file_safety_record,
+    cmd_file_safety_status,
+    cmd_file_safety_unlock,
+)
+from synapse.commands.history import (
+    cmd_history_cleanup,
+    cmd_history_export,
+    cmd_history_list,
+    cmd_history_search,
+    cmd_history_show,
+    cmd_history_stats,
+    cmd_trace,
+)
 from synapse.commands.list import ListCommand
+from synapse.commands.memory import (
+    cmd_memory_delete,
+    cmd_memory_list,
+    cmd_memory_save,
+    cmd_memory_search,
+    cmd_memory_show,
+    cmd_memory_stats,
+)
+from synapse.commands.messaging import (
+    cmd_broadcast,
+    cmd_interrupt,
+    cmd_reply,
+    cmd_send,
+)
 from synapse.commands.session import (
     cmd_session_delete,
     cmd_session_list,
@@ -38,6 +79,12 @@ from synapse.commands.session import (
     cmd_session_save,
     cmd_session_sessions,
     cmd_session_show,
+)
+from synapse.commands.spawn_cmd import (
+    _extract_tool_args,
+    _warn_synapse_flags_in_tool_args,
+    cmd_spawn,
+    cmd_team_start,
 )
 from synapse.commands.start import StartCommand
 from synapse.commands.workflow import (
@@ -62,6 +109,17 @@ from synapse.utils import resolve_command_path
 # Known profiles (for shortcut detection)
 KNOWN_PROFILES = set(PORT_RANGES.keys())
 logger = logging.getLogger(__name__)
+
+# Compatibility aliases used by tests and by command modules that still
+# reference helper functions via ``synapse.cli``.
+_get_history_manager = history_commands._get_history_manager
+_memory_broadcast_notify = memory_commands._memory_broadcast_notify
+_build_a2a_cmd = messaging_commands._build_a2a_cmd
+_get_a2a_tool_path = messaging_commands._get_a2a_tool_path
+_get_send_message_threshold = messaging_commands._get_send_message_threshold
+_resolve_cli_message = messaging_commands._resolve_cli_message
+_resolve_task_message = messaging_commands._resolve_task_message
+_run_a2a_command = messaging_commands._run_a2a_command
 
 
 def install_skills() -> None:
@@ -893,611 +951,6 @@ def _maybe_prompt_save_agent_profile(
 
 
 # ============================================================
-# History Commands
-# ============================================================
-
-
-HISTORY_DISABLED_MSG = "History is disabled. Enable with: SYNAPSE_HISTORY_ENABLED=true"
-FILE_SAFETY_DISABLED_MSG = (
-    "File safety is disabled. Enable with: SYNAPSE_FILE_SAFETY_ENABLED=true"
-)
-
-
-def _get_history_manager() -> HistoryManager:
-    """Get HistoryManager with settings env applied.
-
-    Applies env settings from .synapse/settings.json before checking
-    SYNAPSE_HISTORY_ENABLED environment variable.
-    """
-    from synapse.history import HistoryManager
-    from synapse.settings import get_settings
-
-    # Apply settings env to os.environ
-    settings = get_settings()
-    env_dict = dict(os.environ)
-    settings.apply_env(env_dict)
-    os.environ.update(env_dict)
-
-    from synapse.paths import get_history_db_path
-
-    db_path = get_history_db_path()
-    return HistoryManager.from_env(db_path=db_path)
-
-
-def _print_history_table(observations: list[dict]) -> None:
-    """Print observations in a formatted table."""
-    hdr = f"{'Task ID':<36} {'Agent':<10} {'Status':<12} {'Timestamp':<19}"
-    print(f"{hdr} {'Input (first 40 chars)':<42}")
-    print("-" * 119)
-
-    for obs in observations:
-        task_id = obs["task_id"][:36]
-        agent = obs["agent_name"][:10]
-        status = obs["status"][:12]
-        timestamp = obs["timestamp"][:19] if obs["timestamp"] else "N/A"
-        input_preview = (
-            obs["input"][:40].replace("\n", " ") if obs["input"] else "(empty)"
-        )
-        row = f"{task_id:<36} {agent:<10} {status:<12} {timestamp:<19}"
-        print(f"{row} {input_preview:<42}")
-
-
-def cmd_history_list(args: argparse.Namespace) -> None:
-    """List task history."""
-    manager = _get_history_manager()
-
-    if not manager.enabled:
-        print(HISTORY_DISABLED_MSG)
-        return
-
-    observations = manager.list_observations(
-        limit=args.limit, agent_name=args.agent if args.agent else None
-    )
-
-    if not observations:
-        print("No task history found.")
-        return
-
-    _print_history_table(observations)
-
-    print(f"\nShowing {len(observations)} entries (limit: {args.limit})")
-    if args.agent:
-        print(f"Filtered by agent: {args.agent}")
-
-
-def _print_observation_detail(observation: dict) -> None:
-    """Print detailed observation information (shared by history show and trace)."""
-    import json
-
-    print(f"Task ID:        {observation['task_id']}")
-    print(f"Agent:          {observation['agent_name']}")
-    print(f"Status:         {observation['status']}")
-    print(f"Session ID:     {observation['session_id']}")
-    print(f"Timestamp:      {observation['timestamp']}")
-
-    print("\n" + "=" * 80)
-    print("INPUT:")
-    print("=" * 80)
-    print(observation["input"] or "(empty)")
-
-    print("\n" + "=" * 80)
-    print("OUTPUT:")
-    print("=" * 80)
-    print(observation["output"] or "(empty)")
-
-    if observation.get("metadata"):
-        print("\n" + "=" * 80)
-        print("METADATA:")
-        print("=" * 80)
-        print(json.dumps(observation["metadata"], indent=2))
-
-
-def cmd_history_show(args: argparse.Namespace) -> None:
-    """Show detailed task information."""
-    manager = _get_history_manager()
-
-    if not manager.enabled:
-        print(HISTORY_DISABLED_MSG)
-        return
-
-    observation = manager.get_observation(args.task_id)
-
-    if not observation:
-        print(f"Task not found: {args.task_id}")
-        sys.exit(1)
-
-    _print_observation_detail(observation)
-
-
-def cmd_trace(args: argparse.Namespace) -> None:
-    """Trace a task ID across A2A history and file-safety modification records."""
-    manager = _get_history_manager()
-
-    if not manager.enabled:
-        print(HISTORY_DISABLED_MSG)
-        return
-
-    observation = manager.get_observation(args.task_id)
-    if not observation:
-        print(f"Task not found: {args.task_id}")
-        sys.exit(1)
-
-    _print_observation_detail(observation)
-
-    # File-safety records for the same task_id (best-effort)
-    from synapse.file_safety import FileSafetyManager
-
-    fm = FileSafetyManager.from_env()
-    if not fm.enabled:
-        return
-
-    mods = fm.get_modifications_by_task(observation["task_id"])
-    if not mods:
-        return
-
-    print("\n" + "=" * 80)
-    print("FILE MODIFICATIONS:")
-    print("=" * 80)
-    for mod in mods:
-        ts = mod.get("timestamp", "N/A")
-        agent = mod.get("agent_name", "unknown")
-        change_type = mod.get("change_type", "MODIFY")
-        path = mod.get("file_path", "")
-        print(f"\n[{ts}] {agent} - {change_type}")
-        if mod.get("intent"):
-            print(f"  Intent: {mod['intent']}")
-        if path:
-            print(f"  File: {path}")
-
-
-def cmd_history_search(args: argparse.Namespace) -> None:
-    """Search task history by keywords."""
-    manager = _get_history_manager()
-
-    if not manager.enabled:
-        print(HISTORY_DISABLED_MSG)
-        return
-
-    observations = manager.search_observations(
-        keywords=args.keywords,
-        logic=args.logic,
-        case_sensitive=args.case_sensitive,
-        limit=args.limit,
-        agent_name=args.agent if args.agent else None,
-    )
-
-    if not observations:
-        print(f"No matches found for: {', '.join(args.keywords)}")
-        return
-
-    _print_history_table(observations)
-
-    print(f"\nFound {len(observations)} matches")
-    print(f"Keywords: {', '.join(args.keywords)} (logic: {args.logic})")
-    if args.agent:
-        print(f"Filtered by agent: {args.agent}")
-
-
-def cmd_history_cleanup(args: argparse.Namespace) -> None:
-    """Clean up old task history."""
-    import sqlite3
-
-    manager = _get_history_manager()
-
-    if not manager.enabled:
-        print(HISTORY_DISABLED_MSG)
-        return
-
-    # Validate arguments
-    if args.days is None and args.max_size is None:
-        print("Error: Specify --days or --max-size")
-        sys.exit(1)
-
-    if args.days is not None and args.max_size is not None:
-        print("Error: Specify only one of --days or --max-size")
-        sys.exit(1)
-
-    db_path = manager.db_path
-
-    # Dry-run mode: Show what would be deleted
-    if args.dry_run:
-        if args.days:
-            # Count observations older than N days
-            try:
-                conn = sqlite3.connect(db_path)
-                cursor = conn.cursor()
-                cutoff_sql = f"datetime('now', '-{args.days} days')"
-                cursor.execute(
-                    f"SELECT COUNT(*) FROM observations WHERE timestamp < {cutoff_sql}"
-                )
-                count = cursor.fetchone()[0]
-                conn.close()
-                print(f"Would delete {count} observations older than {args.days} days")
-            except Exception as e:
-                print(f"Error checking observations: {e}", file=sys.stderr)
-        else:
-            # Show current size and target
-            try:
-                current_size_mb = Path(db_path).stat().st_size / (1024 * 1024)
-                print(f"Current database size: {current_size_mb:.2f} MB")
-                print(f"Target size: {args.max_size} MB")
-                if current_size_mb > args.max_size:
-                    print("Would delete oldest observations to reach target size")
-                else:
-                    print("No cleanup needed (already under target size)")
-            except Exception as e:
-                print(f"Error checking database: {e}", file=sys.stderr)
-        return
-
-    # Confirm deletion (unless --force flag)
-    if not args.force:
-        response = input(
-            "This will permanently delete observations. Continue? (yes/no): "
-        )
-        if response.lower() not in ("yes", "y"):
-            print("Cancelled.")
-            return
-
-    # Execute cleanup
-    print("Cleaning up...")
-    if args.days:
-        result = manager.cleanup_old_observations(
-            days=args.days,
-            vacuum=not args.no_vacuum,
-        )
-        count = result["deleted_count"]
-        print(f"Deleted {count} observations older than {args.days} days")
-    else:
-        result = manager.cleanup_by_size(
-            max_size_mb=args.max_size,
-            vacuum=not args.no_vacuum,
-        )
-        print(f"Deleted {result['deleted_count']} observations to reach target size")
-
-    if not args.no_vacuum and result["vacuum_reclaimed_mb"] > 0:
-        print(f"Reclaimed {result['vacuum_reclaimed_mb']:.2f} MB of disk space")
-
-
-def cmd_history_stats(args: argparse.Namespace) -> None:
-    """Show task history statistics."""
-    manager = _get_history_manager()
-
-    if not manager.enabled:
-        print(HISTORY_DISABLED_MSG)
-        return
-
-    stats = manager.get_statistics(agent_name=args.agent if args.agent else None)
-
-    if not stats or stats["total_tasks"] == 0:
-        print("No task history found.")
-        return
-
-    # Print statistics (section-based format)
-    print("=" * 60)
-    print("TASK HISTORY STATISTICS")
-    print("=" * 60)
-    print()
-
-    # Overall metrics
-    print(f"Total Tasks:     {stats['total_tasks']}")
-    print(f"Completed:       {stats['completed']}")
-    print(f"Failed:          {stats['failed']}")
-    print(f"Canceled:        {stats['canceled']}")
-    print(f"Success Rate:    {stats['success_rate']:.1f}%")
-    print()
-
-    # Database info
-    print(f"Database Size:   {stats['db_size_mb']:.2f} MB")
-    if stats["oldest_task"]:
-        print(f"Oldest Task:     {stats['oldest_task']}")
-        print(f"Newest Task:     {stats['newest_task']}")
-        print(f"Date Range:      {stats['date_range_days']} days")
-    print()
-
-    # Agent breakdown (if not filtering by specific agent)
-    if stats["by_agent"]:
-        print("=" * 60)
-        print("BY AGENT")
-        print("=" * 60)
-        print()
-        hdr = f"{'Agent':<10} {'Total':<8} {'Completed':<10} {'Failed':<8}"
-        print(f"{hdr} {'Canceled':<8}")
-        print("-" * 60)
-
-        for agent, counts in sorted(stats["by_agent"].items()):
-            print(
-                f"{agent:<10} {counts['total']:<8} {counts['completed']:<10} "
-                f"{counts['failed']:<8} {counts['canceled']:<8}"
-            )
-        print()
-
-    # Token usage section (non-blocking — never break stats display)
-    try:
-        token_stats = manager.get_token_statistics(
-            agent_name=args.agent if args.agent else None
-        )
-        total_in = token_stats.get("total_input_tokens", 0)
-        total_out = token_stats.get("total_output_tokens", 0)
-        if total_in > 0 or total_out > 0:
-            print("=" * 60)
-            print("TOKEN USAGE")
-            print("=" * 60)
-            print()
-            print(f"Input Tokens:    {total_in:,}")
-            print(f"Output Tokens:   {total_out:,}")
-            total_cost = token_stats.get("total_cost_usd", 0.0)
-            print(f"Total Cost:      ${total_cost:.4f}")
-            print()
-
-            by_agent = token_stats.get("by_agent")
-            if by_agent:
-                hdr = f"{'Agent':<10} {'Input':<12} {'Output':<12} {'Cost':<10}"
-                print(hdr)
-                print("-" * 44)
-                for agent, counts in sorted(by_agent.items()):
-                    print(
-                        f"{agent:<10} {counts.get('input_tokens', 0):<12,} "
-                        f"{counts.get('output_tokens', 0):<12,} "
-                        f"${counts.get('cost_usd', 0.0):<9.4f}"
-                    )
-                print()
-    except Exception:
-        pass  # Token stats are optional; never break the main stats display
-
-    if args.agent:
-        print(f"(Filtered by agent: {args.agent})")
-
-
-def cmd_history_export(args: argparse.Namespace) -> None:
-    """Export task history in specified format."""
-    manager = _get_history_manager()
-
-    if not manager.enabled:
-        print(HISTORY_DISABLED_MSG)
-        return
-
-    # Get export format
-    export_format = args.format.lower()
-    if export_format not in ("json", "csv"):
-        print(f"Error: Invalid format '{export_format}'. Use 'json' or 'csv'.")
-        sys.exit(1)
-
-    # Export with optional filters
-    exported_data = manager.export_observations(
-        format=export_format,
-        agent_name=args.agent if args.agent else None,
-        limit=args.limit if args.limit else None,
-    )
-
-    # Output to file or stdout
-    if args.output:
-        try:
-            with open(args.output, "w") as f:
-                f.write(exported_data)
-            print(f"Exported to {args.output}")
-        except OSError as e:
-            print(f"Error writing to file: {e}", file=sys.stderr)
-            sys.exit(1)
-    else:
-        # Output to stdout
-        print(exported_data)
-
-
-def _get_a2a_tool_path() -> Path:
-    """Get the path to the a2a.py tool from installed package."""
-    import synapse
-
-    return Path(synapse.__file__).parent / "tools" / "a2a.py"
-
-
-def _run_a2a_command(
-    cmd: list[str],
-    exit_on_error: bool = False,
-) -> None:
-    """Run an a2a.py command and print output."""
-    result = subprocess.run(cmd, capture_output=True, text=True)
-    if result.stdout:
-        print(result.stdout, end="")
-    if result.stderr:
-        print(result.stderr, end="", file=sys.stderr)
-    if exit_on_error and result.returncode != 0:
-        sys.exit(result.returncode)
-
-
-def _get_send_message_threshold() -> int:
-    """Get the byte-size threshold for auto temp-file fallback.
-
-    Default: 102400 (100KB). Override with SYNAPSE_SEND_MESSAGE_THRESHOLD env var.
-    """
-    env_val = os.environ.get("SYNAPSE_SEND_MESSAGE_THRESHOLD")
-    if env_val:
-        return int(env_val)
-    return 102_400
-
-
-def _build_a2a_cmd(
-    subcommand: str,
-    message: str,
-    *,
-    target: str | None = None,
-    priority: int | None = None,
-    sender: str | None = None,
-    response_mode: str | None = None,
-    attachments: list[str] | None = None,
-    force: bool = False,
-) -> list[str]:
-    """Build command arguments for a2a.py subcommand.
-
-    If the message exceeds SYNAPSE_SEND_MESSAGE_THRESHOLD bytes (default 100KB),
-    it is written to a temp file and --message-file is used instead of a
-    positional argument, avoiding ARG_MAX shell limits.
-    """
-    import tempfile
-    import uuid
-
-    cmd = [sys.executable, "-m", "synapse.tools.a2a", subcommand]
-
-    if target:
-        cmd.extend(["--target", target])
-    if priority is not None:
-        cmd.extend(["--priority", str(priority)])
-
-    threshold = _get_send_message_threshold()
-    if len(message.encode("utf-8")) > threshold:
-        # Write to temp file to avoid ARG_MAX
-        send_dir = Path(tempfile.gettempdir()) / "synapse-a2a" / "send-messages"
-        send_dir.mkdir(parents=True, exist_ok=True)
-        temp_file = send_dir / f"send-{uuid.uuid4()}.txt"
-        temp_file.write_text(message, encoding="utf-8")
-        cmd.extend(["--message-file", str(temp_file)])
-    else:
-        cmd.append(message)
-
-    if attachments:
-        for path in attachments:
-            cmd.extend(["--attach", path])
-
-    if sender:
-        cmd.extend(["--from", sender])
-    if response_mode:
-        cmd.append(f"--{response_mode}")
-    if force:
-        cmd.append("--force")
-
-    return cmd
-
-
-def _resolve_cli_message(args: argparse.Namespace) -> str:
-    """Resolve message content from positional arg, --message-file, or --stdin.
-
-    Exactly one source must be specified. Multiple or zero sources cause exit.
-    """
-    sources: list[str] = []
-    if getattr(args, "message", None):
-        sources.append("positional")
-    if getattr(args, "message_file", None):
-        sources.append("--message-file")
-    if getattr(args, "stdin", False):
-        sources.append("--stdin")
-
-    if len(sources) > 1:
-        print(
-            f"Error: Multiple message sources specified: {', '.join(sources)}. "
-            "Use exactly one of: positional argument, --message-file, or --stdin.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    if len(sources) == 0:
-        print(
-            "Error: No message provided. Use a positional argument, "
-            "--message-file PATH, or --stdin.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    if getattr(args, "stdin", False):
-        return sys.stdin.read()
-
-    message_file = getattr(args, "message_file", None)
-    if message_file:
-        if message_file == "-":
-            return sys.stdin.read()
-        path = Path(message_file)
-        if not path.exists():
-            print(f"Error: Message file not found: {message_file}", file=sys.stderr)
-            sys.exit(1)
-        return path.read_text(encoding="utf-8")
-
-    return str(args.message)
-
-
-def _resolve_task_message(args: argparse.Namespace) -> str | None:
-    """Resolve spawn task content from --task or --task-file."""
-    task = getattr(args, "task", None)
-    task_file = getattr(args, "task_file", None)
-
-    if task is not None:
-        return str(task)
-
-    if task_file:
-        if task_file == "-":
-            return sys.stdin.read()
-        path = Path(task_file)
-        if not path.exists():
-            print(f"Error: Task file not found: {task_file}", file=sys.stderr)
-            sys.exit(1)
-        return path.read_text(encoding="utf-8")
-
-    return None
-
-
-def cmd_send(args: argparse.Namespace) -> None:
-    """Send a message to an agent."""
-    message = _resolve_cli_message(args)
-
-    cmd = _build_a2a_cmd(
-        "send",
-        message,
-        target=args.target,
-        priority=args.priority,
-        sender=getattr(args, "sender", None),
-        response_mode=getattr(args, "response_mode", None),
-        attachments=getattr(args, "attach", None),
-        force=getattr(args, "force", False),
-    )
-    _run_a2a_command(cmd, exit_on_error=True)
-
-
-def cmd_interrupt(args: argparse.Namespace) -> None:
-    """Send a priority-4 interrupt message to an agent (fire-and-forget)."""
-    cmd = _build_a2a_cmd(
-        "send",
-        args.message,
-        target=args.target,
-        priority=4,
-        sender=getattr(args, "sender", None),
-        response_mode="silent",
-        force=getattr(args, "force", False),
-    )
-    _run_a2a_command(cmd, exit_on_error=True)
-
-
-def cmd_broadcast(args: argparse.Namespace) -> None:
-    """Broadcast a message to all agents in current working directory."""
-    message = _resolve_cli_message(args)
-    cmd = _build_a2a_cmd(
-        "broadcast",
-        message,
-        priority=args.priority,
-        sender=getattr(args, "sender", None),
-        response_mode=getattr(args, "response_mode", None),
-        attachments=getattr(args, "attach", None),
-    )
-    _run_a2a_command(cmd, exit_on_error=True)
-
-
-def cmd_reply(args: argparse.Namespace) -> None:
-    """Reply to the last received A2A message using reply tracking."""
-    message = args.message
-    sender = getattr(args, "sender", None)
-    to_sender = getattr(args, "to", None)
-    list_targets = getattr(args, "list_targets", False)
-
-    cmd = [sys.executable, str(_get_a2a_tool_path()), "reply"]
-    if sender:
-        cmd.extend(["--from", sender])
-    if to_sender:
-        cmd.extend(["--to", to_sender])
-    if list_targets:
-        cmd.append("--list-targets")
-    if message:
-        cmd.append(message)
-
-    _run_a2a_command(cmd, exit_on_error=True)
-
-
-# ============================================================
 # Instructions Commands
 # ============================================================
 
@@ -1599,591 +1052,6 @@ def cmd_mcp_serve(args: argparse.Namespace) -> None:
         port=args.port,
     )
     serve_stdio(server)
-
-
-# ============================================================
-# File Safety Commands
-# ============================================================
-
-
-def cmd_file_safety_status(args: argparse.Namespace) -> None:
-    """Show file safety statistics."""
-    from synapse.file_safety import FileSafetyManager
-
-    manager = FileSafetyManager.from_env()
-
-    if not manager.enabled:
-        print(FILE_SAFETY_DISABLED_MSG)
-        return
-
-    stats = manager.get_statistics()
-
-    if not stats:
-        print("No file safety data found.")
-        return
-
-    print("=" * 60)
-    print("FILE SAFETY STATISTICS")
-    print("=" * 60)
-    print()
-    print(f"Active Locks:        {stats.get('active_locks', 0)}")
-    print(f"Total Modifications: {stats.get('total_modifications', 0)}")
-    print()
-
-    by_type = stats.get("by_change_type", {})
-    if by_type:
-        print("By Change Type:")
-        for change_type, count in by_type.items():
-            print(f"  {change_type}: {count}")
-        print()
-
-    by_agent = stats.get("by_agent", {})
-    if by_agent:
-        print("By Agent:")
-        for agent, count in by_agent.items():
-            print(f"  {agent}: {count}")
-        print()
-
-    most_modified = stats.get("most_modified_files", [])
-    if most_modified:
-        print("Most Modified Files:")
-        for item in most_modified[:5]:
-            print(f"  {item['file_path']}: {item['count']} modifications")
-
-
-def cmd_file_safety_locks(args: argparse.Namespace) -> None:
-    """List active file locks."""
-    import os
-
-    from synapse.file_safety import FileSafetyManager
-
-    manager = FileSafetyManager.from_env()
-
-    if not manager.enabled:
-        print(FILE_SAFETY_DISABLED_MSG)
-        return
-
-    # Get locks with stale information
-    locks = manager.list_locks(
-        agent_name=getattr(args, "agent", None),
-        agent_type=getattr(args, "type", None),
-        include_stale=True,
-    )
-
-    if not locks:
-        print("No active file locks.")
-        return
-
-    # Check for stale locks
-    stale_count = 0
-    live_count = 0
-
-    print(
-        f"{'File Path':<40} {'Agent':<20} {'PID':<8} {'Status':<8} {'Expires At':<20}"
-    )
-    print("-" * 100)
-
-    for lock in locks:
-        file_path = os.path.basename(lock["file_path"])[:40]
-        agent = (lock.get("agent_id") or lock.get("agent_name", "unknown"))[:20]
-        pid = lock.get("pid")
-        expires = lock["expires_at"][:20] if lock.get("expires_at") else "N/A"
-
-        # Determine if lock is stale
-        if pid:
-            is_alive = manager._is_process_running(pid)
-            status = "LIVE" if is_alive else "STALE"
-            if is_alive:
-                live_count += 1
-            else:
-                stale_count += 1
-        else:
-            status = "UNKNOWN"
-            live_count += 1  # Assume old locks without PID are live
-
-        pid_str = str(pid) if pid else "-"
-        print(f"{file_path:<40} {agent:<20} {pid_str:<8} {status:<8} {expires:<20}")
-
-    print(f"\nTotal: {len(locks)} locks ({live_count} live, {stale_count} stale)")
-
-    if stale_count > 0:
-        print(f"\nWarning: {stale_count} stale lock(s) from dead processes detected.")
-        print("Run 'synapse file-safety cleanup-locks' to clean them up.")
-
-
-def _resolve_agent_info(
-    agent_name: str,
-) -> tuple[str, str | None, int | None]:
-    """Resolve agent name to (agent_id, agent_type, pid).
-
-    Lookup order:
-    1. Exact match by agent_id in registry
-    2. Match by agent_type (short name like "claude") in live agents
-    3. Parse agent_type from agent_id format "synapse-{type}-{port}"
-
-    Returns:
-        Tuple of (agent_id, agent_type, pid). agent_type and pid may be None.
-    """
-    from synapse.registry import AgentRegistry
-
-    agent_id = agent_name
-    agent_type = None
-    pid = None
-
-    with contextlib.suppress(Exception):
-        registry = AgentRegistry()
-        agent_info = registry.get_agent(agent_id)
-
-        # Try to find by agent_type if exact match fails
-        if not agent_info:
-            for aid, info in registry.get_live_agents().items():
-                if info.get("agent_type") == agent_id:
-                    agent_info = info
-                    agent_id = aid
-                    break
-
-        if agent_info:
-            pid = agent_info.get("pid")
-            agent_type = agent_info.get("agent_type")
-
-    # Fallback: extract agent_type from agent_id format "synapse-{type}-{port}"
-    if not agent_type and agent_id.startswith("synapse-"):
-        parts = agent_id.split("-")
-        if len(parts) >= 3:
-            agent_type = parts[1]
-
-    return agent_id, agent_type, pid
-
-
-def cmd_file_safety_lock(args: argparse.Namespace) -> None:
-    """Acquire a lock on a file."""
-    from synapse.file_safety import FileSafetyManager, LockStatus
-
-    manager = FileSafetyManager.from_env()
-
-    if not manager.enabled:
-        print(FILE_SAFETY_DISABLED_MSG)
-        return
-
-    agent_id, agent_type, pid = _resolve_agent_info(args.agent)
-    wait = getattr(args, "wait", False)
-    wait_timeout = getattr(args, "wait_timeout", None)
-    wait_interval = getattr(args, "wait_interval", 2.0)
-    start_time = time.monotonic()
-
-    while True:
-        result = manager.acquire_lock(
-            file_path=args.file,
-            agent_id=agent_id,
-            agent_type=agent_type,
-            task_id=getattr(args, "task_id", None),
-            duration_seconds=getattr(args, "duration", None),
-            intent=getattr(args, "intent", None),
-            pid=pid,
-        )
-
-        status = result["status"]
-        if status == LockStatus.ACQUIRED:
-            print(f"Lock acquired on {args.file}")
-            print(f"Expires at: {result.get('expires_at')}")
-            return
-        if status == LockStatus.RENEWED:
-            print(f"Lock renewed on {args.file}")
-            print(f"New expiration: {result.get('expires_at')}")
-            return
-        if status == LockStatus.ALREADY_LOCKED:
-            if not wait:
-                print(
-                    f"File is already locked by {result.get('lock_holder')}",
-                    file=sys.stderr,
-                )
-                print(f"Expires at: {result.get('expires_at')}", file=sys.stderr)
-                sys.exit(1)
-            # Check timeout before sleeping
-            elapsed = time.monotonic() - start_time
-            if wait_timeout is not None and elapsed >= wait_timeout:
-                print(f"Timed out waiting for lock on {args.file}")
-                sys.exit(1)
-            lock_holder = result.get("lock_holder")
-            if lock_holder:
-                print(f"Waiting for lock on {args.file} (held by {lock_holder})...")
-            else:
-                print(f"Waiting for lock on {args.file}...")
-            if result.get("expires_at"):
-                print(f"Current lock expires at: {result.get('expires_at')}")
-            # Clamp sleep to remaining time to avoid oversleeping past timeout
-            if wait_timeout is not None:
-                remaining = wait_timeout - elapsed
-                sleep_time = min(float(wait_interval), max(0.0, remaining))
-            else:
-                sleep_time = max(0.0, float(wait_interval))
-            time.sleep(sleep_time)
-            continue
-        if status == LockStatus.FAILED:
-            print(
-                f"Failed to acquire lock: {result.get('error', 'unknown error')}",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-        print(f"Unexpected lock status: {status}", file=sys.stderr)
-        sys.exit(1)
-
-
-def cmd_file_safety_unlock(args: argparse.Namespace) -> None:
-    """Release a lock on a file."""
-    from synapse.file_safety import FileSafetyManager
-
-    manager = FileSafetyManager.from_env()
-
-    if not manager.enabled:
-        print(FILE_SAFETY_DISABLED_MSG)
-        return
-
-    force = getattr(args, "force", False)
-    agent = getattr(args, "agent", None)
-
-    if not force and not agent:
-        print("Error: Agent name is required unless using --force", file=sys.stderr)
-        sys.exit(1)
-
-    if force:
-        # Force unlock regardless of owner
-        if manager.force_unlock(args.file):
-            print(f"Lock force-released on {args.file}")
-        else:
-            print(f"No lock found for {args.file}")
-            sys.exit(1)
-    else:
-        # Normal unlock (requires agent ownership)
-        # agent is guaranteed to be str here due to earlier check
-        if manager.release_lock(args.file, str(agent)):
-            print(f"Lock released on {args.file}")
-        else:
-            print(f"No lock found for {args.file} by {agent}")
-            sys.exit(1)
-
-
-def cmd_file_safety_cleanup_locks(args: argparse.Namespace) -> None:
-    """Clean up stale locks from dead processes."""
-    from synapse.file_safety import FileSafetyManager
-
-    manager = FileSafetyManager.from_env()
-
-    if not manager.enabled:
-        print(FILE_SAFETY_DISABLED_MSG)
-        return
-
-    # First show what will be cleaned
-    stale_locks = manager.get_stale_locks()
-
-    if not stale_locks:
-        print("No stale locks found.")
-        return
-
-    print(f"Found {len(stale_locks)} stale lock(s) from dead processes:")
-    for lock in stale_locks:
-        pid = lock.get("pid")
-        agent = lock.get("agent_id")
-        print(f"  - {lock['file_path']} (pid={pid}, agent={agent})")
-
-    # Check for --force or prompt for confirmation
-    force = getattr(args, "force", False)
-    if not force:
-        response = input("\nClean up these locks? [y/N]: ").strip().lower()
-        if response != "y":
-            print("Aborted.")
-            return
-
-    # Clean up
-    cleaned = manager.cleanup_stale_locks()
-    print(f"\nCleaned up {cleaned} stale lock(s).")
-
-
-def cmd_file_safety_history(args: argparse.Namespace) -> None:
-    """Show modification history for a file."""
-    from synapse.file_safety import FileSafetyManager
-
-    manager = FileSafetyManager.from_env()
-
-    if not manager.enabled:
-        print(FILE_SAFETY_DISABLED_MSG)
-        return
-
-    history = manager.get_file_history(args.file, limit=args.limit)
-
-    if not history:
-        print(f"No modification history found for {args.file}")
-        return
-
-    print(f"Modification history for: {args.file}")
-    print("=" * 80)
-
-    for mod in history:
-        print(f"\n[{mod['timestamp']}] {mod['agent_name']} - {mod['change_type']}")
-        if mod.get("intent"):
-            print(f"  Intent: {mod['intent']}")
-        if mod.get("affected_lines"):
-            print(f"  Lines: {mod['affected_lines']}")
-        print(f"  Task ID: {mod['task_id']}")
-
-
-def cmd_file_safety_recent(args: argparse.Namespace) -> None:
-    """Show recent file modifications."""
-    from synapse.file_safety import FileSafetyManager
-
-    manager = FileSafetyManager.from_env()
-
-    if not manager.enabled:
-        print(FILE_SAFETY_DISABLED_MSG)
-        return
-
-    mods = manager.get_recent_modifications(
-        limit=args.limit,
-        agent_name=getattr(args, "agent", None),
-    )
-
-    if not mods:
-        print("No recent modifications found.")
-        return
-
-    print(f"{'Timestamp':<20} {'Agent':<12} {'Type':<8} {'File':<40}")
-    print("-" * 80)
-
-    for mod in mods:
-        timestamp = mod["timestamp"][:20] if mod.get("timestamp") else "N/A"
-        agent = mod["agent_name"][:12]
-        change_type = mod["change_type"][:8]
-        file_path = mod["file_path"][-40:]  # Show last 40 chars
-        print(f"{timestamp:<20} {agent:<12} {change_type:<8} {file_path:<40}")
-
-    print(f"\nShowing {len(mods)} recent modifications")
-
-
-def cmd_file_safety_record(args: argparse.Namespace) -> None:
-    """Record a file modification."""
-    from synapse.file_safety import FileSafetyManager
-
-    manager = FileSafetyManager.from_env()
-
-    if not manager.enabled:
-        print(FILE_SAFETY_DISABLED_MSG)
-        return
-
-    record_id = manager.record_modification(
-        file_path=args.file_path,
-        agent_name=args.agent,
-        task_id=args.task_id,
-        change_type=args.type,
-        intent=args.intent,
-    )
-
-    if record_id:
-        print(f"Recorded modification (ID: {record_id})")
-        print(f"  File: {args.file_path}")
-        print(f"  Agent: {args.agent}")
-        print(f"  Type: {args.type}")
-        if args.intent:
-            print(f"  Intent: {args.intent}")
-    else:
-        print("Failed to record modification")
-
-
-def cmd_file_safety_cleanup(args: argparse.Namespace) -> None:
-    """Clean up old modification records."""
-    from synapse.file_safety import FileSafetyManager
-
-    manager = FileSafetyManager.from_env()
-
-    if not manager.enabled:
-        print(FILE_SAFETY_DISABLED_MSG)
-        return
-
-    # Confirm
-    if not args.force:
-        response = input(
-            f"Delete modification records older than {args.days} days? (yes/no): "
-        )
-        if response.lower() not in ("yes", "y"):
-            print("Cancelled.")
-            return
-
-    deleted = manager.cleanup_old_modifications(days=args.days)
-    print(f"Deleted {deleted} modification records older than {args.days} days")
-
-    # Also cleanup expired locks
-    expired_locks = manager.cleanup_expired_locks()
-    if expired_locks > 0:
-        print(f"Cleaned up {expired_locks} expired locks")
-
-
-def cmd_file_safety_debug(args: argparse.Namespace) -> None:
-    """Show debug information for file safety troubleshooting."""
-    from synapse.file_safety import FileSafetyManager
-
-    print("=" * 60)
-    print("FILE SAFETY DEBUG INFORMATION")
-    print("=" * 60)
-
-    # Environment variables
-    print("\n[Environment Variables]")
-    env_vars = [
-        "SYNAPSE_FILE_SAFETY_ENABLED",
-        "SYNAPSE_FILE_SAFETY_RETENTION_DAYS",
-        "SYNAPSE_LOG_LEVEL",
-    ]
-    for var in env_vars:
-        value = os.environ.get(var, "(not set)")
-        print(f"  {var}: {value}")
-
-    # Settings file locations
-    print("\n[Settings Files]")
-    settings_paths = [
-        Path.cwd() / ".synapse" / "settings.json",
-        Path.home() / ".synapse" / "settings.json",
-    ]
-    for path in settings_paths:
-        status = "exists" if path.exists() else "not found"
-        print(f"  {path}: {status}")
-
-    # Database info
-    print("\n[Database]")
-    manager = FileSafetyManager.from_env()
-    print(f"  Enabled: {manager.enabled}")
-    print(f"  DB Path: {manager.db_path}")
-    print(f"  DB Exists: {Path(manager.db_path).exists()}")
-    print(f"  Retention Days: {manager.retention_days}")
-
-    if manager.enabled:
-        stats = manager.get_statistics()
-        print(f"  Active Locks: {stats['active_locks']}")
-        print(f"  Total Modifications: {stats['total_modifications']}")
-
-    # Instruction files
-    print("\n[Instruction Files]")
-    instruction_files = [
-        Path.cwd() / ".synapse" / "file-safety.md",
-        Path.home() / ".synapse" / "file-safety.md",
-    ]
-    for path in instruction_files:
-        status = "exists" if path.exists() else "not found"
-        print(f"  {path}: {status}")
-
-    # Log file locations
-    print("\n[Log Files]")
-    log_dir = Path.home() / ".synapse" / "logs"
-    if log_dir.exists():
-        log_files = list(log_dir.glob("*.log"))
-        if log_files:
-            for lf in sorted(log_files)[-5:]:  # Show last 5
-                print(f"  {lf}")
-        else:
-            print("  (no log files)")
-    else:
-        print(f"  Log directory not found: {log_dir}")
-
-    print("\n[Debug Tips]")
-    print("  - Enable debug logging: SYNAPSE_LOG_LEVEL=DEBUG synapse ...")
-    print("  - Enable file logging: SYNAPSE_LOG_FILE=true synapse ...")
-    print("  - View active locks: synapse file-safety locks")
-    print("  - View recent mods: synapse file-safety recent")
-
-
-# ============================================================
-# External Agent Management Commands
-# ============================================================
-
-
-def cmd_external_add(args: argparse.Namespace) -> None:
-    """Add an external A2A agent."""
-    client = get_client()
-    agent = client.discover(args.url, alias=args.alias)
-
-    if agent:
-        print(f"Added external agent: {agent.name}")
-        print(f"  Alias: {agent.alias}")
-        print(f"  URL: {agent.url}")
-        print(f"  Description: {agent.description}")
-        if agent.skills:
-            skills_str = ", ".join(s.get("name", s.get("id", "")) for s in agent.skills)
-            print(f"  Skills: {skills_str}")
-    else:
-        print(f"Failed to add agent from {args.url}")
-        sys.exit(1)
-
-
-def cmd_external_list(args: argparse.Namespace) -> None:
-    """List external A2A agents."""
-    client = get_client()
-    agents = client.list_agents()
-
-    if not agents:
-        print("No external agents registered.")
-        print("Use 'synapse external add <url>' to add one.")
-        return
-
-    print(f"{'ALIAS':<15} {'NAME':<20} {'URL':<40} {'LAST SEEN'}")
-    print("-" * 90)
-    for agent in agents:
-        last_seen = agent.last_seen[:19] if agent.last_seen else "Never"
-        print(f"{agent.alias:<15} {agent.name:<20} {agent.url:<40} {last_seen}")
-
-
-def cmd_external_remove(args: argparse.Namespace) -> None:
-    """Remove an external A2A agent."""
-    client = get_client()
-
-    if client.remove_agent(args.alias):
-        print(f"Removed external agent: {args.alias}")
-    else:
-        print(f"Agent '{args.alias}' not found")
-        sys.exit(1)
-
-
-def cmd_external_send(args: argparse.Namespace) -> None:
-    """Send a message to an external A2A agent."""
-    client = get_client()
-
-    task = client.send_message(args.alias, args.message, wait_for_completion=args.wait)
-
-    if task:
-        print(f"Task ID: {task.id}")
-        print(f"Status: {task.status}")
-        if task.artifacts:
-            print("Artifacts:")
-            for artifact in task.artifacts:
-                print(f"  - {artifact}")
-    else:
-        print(f"Failed to send message to {args.alias}")
-        sys.exit(1)
-
-
-def cmd_external_info(args: argparse.Namespace) -> None:
-    """Show detailed info about an external agent."""
-    client = get_client()
-    agent = client.registry.get(args.alias)
-
-    if not agent:
-        print(f"Agent '{args.alias}' not found")
-        sys.exit(1)
-
-    print(f"Name: {agent.name}")
-    print(f"Alias: {agent.alias}")
-    print(f"URL: {agent.url}")
-    print(f"Description: {agent.description}")
-    print(f"Added: {agent.added_at}")
-    print(f"Last Seen: {agent.last_seen or 'Never'}")
-
-    if agent.capabilities:
-        print("\nCapabilities:")
-        for key, value in agent.capabilities.items():
-            print(f"  {key}: {value}")
-
-    if agent.skills:
-        print("\nSkills:")
-        for skill in agent.skills:
-            print(f"  - {skill.get('name', skill.get('id', 'Unknown'))}")
-            if skill.get("description"):
-                print(f"    {skill['description']}")
 
 
 # ============================================================
@@ -3122,574 +1990,6 @@ def interactive_agent_setup(
         if original_settings is not None and termios is not None:
             with contextlib.suppress(termios.error):
                 termios.tcsetattr(sys.stdin, termios.TCSADRAIN, original_settings)
-
-
-def cmd_memory_save(args: argparse.Namespace) -> None:
-    """Save a memory entry."""
-    from synapse.shared_memory import SharedMemory
-
-    mem = SharedMemory.from_env()
-    tags = (
-        [t.strip() for t in args.tags.split(",") if t.strip()]
-        if getattr(args, "tags", None)
-        else None
-    )
-    author = os.environ.get("SYNAPSE_AGENT_ID", "user")
-
-    result = mem.save(
-        key=args.key,
-        content=args.content,
-        author=author,
-        tags=tags,
-        scope=getattr(args, "scope", "global"),
-    )
-    if result:
-        tag_str = f" [{', '.join(result['tags'])}]" if result["tags"] else ""
-        scope_str = f" ({result['scope']})"
-        print(f"Saved: {result['key']}{tag_str}{scope_str} (id: {result['id'][:8]})")
-    else:
-        print("Shared memory is disabled.", file=sys.stderr)
-        sys.exit(1)
-
-    if getattr(args, "notify", False):
-        _memory_broadcast_notify(args.key)
-
-
-def _memory_broadcast_notify(key: str) -> None:
-    """Broadcast a notification about a saved memory."""
-    try:
-        registry = AgentRegistry()
-        client = A2AClient()
-        agents = registry.list_agents()
-        my_id = os.environ.get("SYNAPSE_AGENT_ID", "")
-        for agent_id, agent_info in agents.items():
-            if agent_id != my_id:
-                name = agent_info.get("name") or agent_id
-                endpoint = agent_info.get("endpoint", "")
-                if not endpoint:
-                    print(f"  Skipped: {name} (no endpoint)")
-                    continue
-                try:
-                    task = client.send_to_local(
-                        endpoint=endpoint,
-                        message=f"Shared memory updated: {key}",
-                        response_mode="silent",
-                    )
-                    if task is None:
-                        print(f"  Failed: {name}: no response")
-                    else:
-                        print(f"  Notified: {name}")
-                except Exception as e:
-                    print(f"  Failed: {name}: {e}")
-    except Exception as e:
-        print(f"  Broadcast failed: {e}", file=sys.stderr)
-
-
-def _resolve_memory_scope(
-    args: argparse.Namespace,
-) -> tuple[str, str | None, str | None]:
-    """Resolve scope, working_dir, and author from CLI args.
-
-    Returns:
-        (scope, working_dir, private_author)
-    """
-    scope = getattr(args, "scope", "global")
-    working_dir = os.getcwd() if scope == "project" else None
-    private_author = (
-        os.environ.get("SYNAPSE_AGENT_ID", "user") if scope == "private" else None
-    )
-    return scope, working_dir, private_author
-
-
-def cmd_memory_list(args: argparse.Namespace) -> None:
-    """List memory entries."""
-    from synapse.shared_memory import SharedMemory
-
-    mem = SharedMemory.from_env()
-    tags = (
-        [t.strip() for t in args.tags.split(",") if t.strip()]
-        if getattr(args, "tags", None)
-        else None
-    )
-    limit = getattr(args, "limit", 50) or 50
-    scope, working_dir, private_author = _resolve_memory_scope(args)
-
-    items = mem.list_memories(
-        author=private_author or getattr(args, "author", None),
-        tags=tags,
-        scope=scope,
-        working_dir=working_dir,
-        limit=limit,
-    )
-
-    if not items:
-        print("No memories found.")
-        return
-
-    for item in items:
-        tag_str = f" [{', '.join(item['tags'])}]" if item["tags"] else ""
-        print(
-            f"  {item['id'][:8]}  {item['key']}{tag_str}"
-            f"  ({item['scope']})  by {item['author']}"
-        )
-
-
-def cmd_memory_show(args: argparse.Namespace) -> None:
-    """Show memory details."""
-    from synapse.shared_memory import SharedMemory
-
-    mem = SharedMemory.from_env()
-    item = mem.get(args.id_or_key)
-
-    if not item:
-        print(f"Memory not found: {args.id_or_key}", file=sys.stderr)
-        sys.exit(1)
-
-    print(f"Key:        {item['key']}")
-    print(f"ID:         {item['id']}")
-    print(f"Author:     {item['author']}")
-    print(f"Scope:      {item['scope']}")
-    if item.get("working_dir"):
-        print(f"WorkingDir: {item['working_dir']}")
-    print(f"Tags:       {', '.join(item['tags']) if item['tags'] else '(none)'}")
-    print(f"Created:    {item['created_at']}")
-    print(f"Updated:    {item['updated_at']}")
-    print(f"\n{item['content']}")
-
-
-def cmd_memory_search(args: argparse.Namespace) -> None:
-    """Search memories."""
-    from synapse.shared_memory import SharedMemory
-
-    mem = SharedMemory.from_env()
-    scope, working_dir, author = _resolve_memory_scope(args)
-    results = mem.search(
-        args.query,
-        scope=scope,
-        author=author,
-        working_dir=working_dir,
-    )
-
-    if not results:
-        print("No matching memories found.")
-        return
-
-    for item in results:
-        tag_str = f" [{', '.join(item['tags'])}]" if item["tags"] else ""
-        print(
-            f"  {item['id'][:8]}  {item['key']}{tag_str}"
-            f"  ({item['scope']})  by {item['author']}"
-        )
-
-
-def cmd_memory_delete(args: argparse.Namespace) -> None:
-    """Delete a memory entry."""
-    from synapse.shared_memory import SharedMemory
-
-    if not getattr(args, "force", False):
-        answer = input(f"Delete memory '{args.id_or_key}'? [y/N] ")
-        if answer.lower() != "y":
-            print("Cancelled.")
-            return
-
-    mem = SharedMemory.from_env()
-    deleted = mem.delete(args.id_or_key)
-
-    if deleted:
-        print(f"Deleted: {args.id_or_key}")
-    else:
-        print(f"Memory not found: {args.id_or_key}", file=sys.stderr)
-        sys.exit(1)
-
-
-def cmd_memory_stats(args: argparse.Namespace) -> None:
-    """Show memory statistics."""
-    from synapse.shared_memory import SharedMemory
-
-    del args  # unused
-    mem = SharedMemory.from_env()
-    stats = mem.stats()
-
-    print(f"Total memories: {stats['total']}")
-
-    if stats["by_author"]:
-        print("\nBy author:")
-        for author, count in sorted(stats["by_author"].items()):
-            print(f"  {author}: {count}")
-
-    if stats["by_tag"]:
-        print("\nBy tag:")
-        for tag, count in sorted(stats["by_tag"].items()):
-            print(f"  {tag}: {count}")
-
-
-# ============================================================
-# Team Commands (B6)
-# ============================================================
-
-
-def _extract_tool_args(items: list[str]) -> tuple[list[str], list[str]]:
-    """Split a list at the first ``--`` separator.
-
-    Used to extract tool_args from positional argument lists where
-    ``argparse.REMAINDER`` can't reliably separate them (e.g., when
-    ``nargs="+"`` consumes the ``--`` token).
-
-    Args:
-        items: List that may contain ``--`` followed by tool arguments.
-
-    Returns:
-        Tuple of (cleaned_items, tool_args).  ``tool_args`` is everything
-        after the first ``--``; ``cleaned_items`` is everything before it.
-    """
-    try:
-        idx = items.index("--")
-        return items[:idx], items[idx + 1 :]
-    except ValueError:
-        return items, []
-
-
-# Known Synapse CLI flags that should NOT appear as tool_args.
-# If these appear after '--', the user likely placed them in the wrong position.
-_SYNAPSE_FLAGS = {
-    "--port",
-    "--name",
-    "-n",
-    "--role",
-    "-r",
-    "--skill-set",
-    "-S",
-    "--terminal",
-    "--foreground",
-    "-f",
-    "--no-setup",
-    "--headless",
-    "--delegate-mode",
-    "--ssl-cert",
-    "--ssl-key",
-    "--layout",
-    "--all-new",
-    "--worktree",
-    # Note: -w omitted to avoid false positives with CLI tools that use -w
-}
-
-
-def _warn_synapse_flags_in_tool_args(tool_args: list[str]) -> bool:
-    """Check if any known Synapse flags appear in tool_args and warn.
-
-    When users accidentally place Synapse flags after the ``--`` separator,
-    those flags are passed to the underlying CLI tool instead of being
-    processed by Synapse.  This function detects that mistake and prints
-    an actionable warning to stderr.
-
-    Args:
-        tool_args: The tool_args list (everything after ``--``).
-
-    Returns:
-        True if a warning was emitted, False otherwise.
-    """
-    if not tool_args:
-        return False
-
-    found = [arg for arg in tool_args if arg in _SYNAPSE_FLAGS]
-    if not found:
-        return False
-
-    flags_str = ", ".join(found)
-    print(
-        f"Warning: Synapse flag(s) {flags_str} found after '--' separator.\n"
-        f"  These will be passed to the underlying CLI tool, not to Synapse.\n"
-        f"  Move them before '--' or before the profile name.\n"
-        f"  Example: synapse spawn --port 8100 claude -- --resume",
-        file=sys.stderr,
-    )
-    return True
-
-
-def _extract_team_agent_name(spec: str) -> str | None:
-    """Extract custom name from team agent spec if present.
-
-    Spec format: profile[:name[:role[:skill_set[:port]]]]
-    """
-    parts = spec.split(":")
-    if len(parts) > 1 and parts[1]:
-        return parts[1]
-    return None
-
-
-def _resolve_team_agent_spec(spec: str, store: AgentProfileStore) -> str:
-    """Resolve team spec target as profile or saved agent ID/name.
-
-    Input spec format: target[:name[:role[:skill_set[:port]]]]
-    Output spec format: profile[:name[:role[:skill_set[:port]]]]
-    """
-    parts = spec.split(":")
-    target = parts[0]
-
-    if target in KNOWN_PROFILES:
-        return spec
-
-    try:
-        saved = store.resolve(target)
-    except AgentProfileError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-    if saved is None:
-        print(
-            f"Error: Unknown profile or saved agent: {target}\n"
-            "Run 'synapse agents list' to see saved agent IDs/names.",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    while len(parts) < 5:
-        parts.append("")
-    parts[0] = saved.profile
-    if not parts[1]:
-        parts[1] = saved.name
-    if not parts[2] and saved.role:
-        parts[2] = saved.role
-    if not parts[3] and saved.skill_set:
-        parts[3] = saved.skill_set
-    return ":".join(parts)
-
-
-def cmd_team_start(args: argparse.Namespace) -> None:
-    """Start multiple agents with split panes.
-
-    Default: First agent replaces current process, others in new panes.
-    --all-new: All agents start in new panes, current process remains.
-
-    Uses prepare_spawn() + execute_spawn() so that all spawn logic
-    (auto-approve flag injection, port allocation, worktree creation)
-    is shared with ``synapse spawn``.
-    """
-    from synapse.spawn import execute_spawn, prepare_spawn
-    from synapse.terminal_jump import detect_terminal_app
-
-    # Strip any '--' + trailing args still embedded in the agents list
-    agents, embedded_tool_args = _extract_tool_args(args.agents)
-    tool_args = getattr(args, "tool_args", []) or embedded_tool_args
-    layout = getattr(args, "layout", "split")
-    all_new = getattr(args, "all_new", False)
-    auto_approve = not getattr(args, "no_auto_approve", False)
-
-    # Resolve each team spec target as profile or saved agent ID/name.
-    store = AgentProfileStore()
-    agents = [_resolve_team_agent_spec(spec, store) for spec in agents]
-
-    # Validate custom names in team specs before launching any process.
-    registry = AgentRegistry()
-    seen_names: set[str] = set()
-    for spec in agents:
-        custom_name = _extract_team_agent_name(spec)
-        if not custom_name:
-            continue
-        if custom_name in seen_names:
-            print(
-                f"Error: Duplicate custom name '{custom_name}' in team specs.",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-        seen_names.add(custom_name)
-        if not registry.is_name_unique(custom_name):
-            print(
-                f"Error: Name '{custom_name}' is already taken by another agent.",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-
-    # Detect terminal early so we can fall back before preparing agents
-    terminal = detect_terminal_app()
-    if not terminal:
-        print(
-            "Could not detect terminal. Supported: tmux, iTerm2, Terminal.app, zellij"
-        )
-        print("Falling back to sequential start...")
-        for agent in agents:
-            print(f"Starting {agent}...")
-            fallback_cmd = ["synapse", agent]
-            if tool_args:
-                fallback_cmd += ["--"] + tool_args
-            subprocess.Popen(
-                fallback_cmd,
-                start_new_session=True,
-            )
-        return
-
-    worktree_opt = getattr(args, "worktree", None)
-    branch_opt = getattr(args, "branch", None)
-
-    # Default to worktree when 2+ agents (unless explicitly opted out)
-    if worktree_opt is None and len(agents) >= 2:
-        worktree_opt = True
-
-    # Prepare all agents via shared spawn infrastructure.
-    # This handles port allocation, worktree creation, and auto-approve
-    # flag injection uniformly.
-    prepared_agents = []
-    for i, agent_spec in enumerate(agents):
-        parts = agent_spec.split(":")
-        profile = parts[0]
-        name = parts[1] if len(parts) > 1 and parts[1] else None
-        role = parts[2] if len(parts) > 2 and parts[2] else None
-        skill_set = parts[3] if len(parts) > 3 and parts[3] else None
-        port = int(parts[4]) if len(parts) > 4 and parts[4] else None
-
-        # Determine worktree name
-        wt: str | bool | None = None
-        if worktree_opt:
-            if isinstance(worktree_opt, str):
-                wt = f"{worktree_opt}-{profile}-{i}"
-            else:
-                wt = True
-
-        try:
-            prepared = prepare_spawn(
-                profile=profile,
-                port=port,
-                name=name,
-                role=role,
-                skill_set=skill_set,
-                tool_args=tool_args or None,
-                worktree=wt,
-                auto_approve=auto_approve,
-                branch=branch_opt,
-            )
-            prepared_agents.append(prepared)
-        except (RuntimeError, ValueError, FileNotFoundError) as e:
-            print(f"Error preparing {profile}: {e}", file=sys.stderr)
-            sys.exit(1)
-
-    if not prepared_agents:
-        print("No agents were prepared.", file=sys.stderr)
-        sys.exit(1)
-
-    # Execute all agents in one batch for proper tile layout.
-    # replace_first=True makes the first agent take over the current terminal
-    # (os.execvp behavior), unless --all-new is specified.
-    try:
-        results = execute_spawn(
-            prepared_agents,
-            layout=layout,
-            terminal=terminal,
-            all_new=all_new,
-            replace_first=not all_new,
-        )
-    except RuntimeError as e:
-        print(f"Error starting agents: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    # Print results
-    if worktree_opt:
-        launched = [r for r in results if r.status == "submitted"]
-        for r in launched:
-            profile = r.agent_id.split("-")[1]
-            print(f"  {profile}: {r.worktree_path} ({r.worktree_branch})")
-        if launched:
-            names = [r.agent_id.split("-")[1] for r in launched]
-            print(f"Started {len(launched)} agents in worktrees: {', '.join(names)}")
-        else:
-            print("No agents were launched.", file=sys.stderr)
-            sys.exit(1)
-    else:
-        display_names = [a.profile for a in prepared_agents]
-        # If replace_first was used, we won't reach this line for the first agent
-        # (os.execvp replaces the process). Print only if all_new.
-        if all_new:
-            print(f"Started {len(prepared_agents)} agents: {', '.join(display_names)}")
-        else:
-            # First agent replaced current process via os.execvp in execute_spawn.
-            # If we reach here, it means replace_first was False or execvp failed.
-            print(f"Handing over terminal to {display_names[0]}...")
-
-
-# ============================================================
-# Spawn Command
-# ============================================================
-
-
-def cmd_spawn(args: argparse.Namespace) -> None:
-    """Spawn a single agent in a new terminal pane."""
-    from synapse.spawn import spawn_agent, wait_for_agent
-
-    tool_args = getattr(args, "tool_args", [])
-    task_message = _resolve_task_message(args)
-    target = args.profile
-    resolved_profile = target
-    resolved_name = args.name
-    resolved_role = args.role
-    resolved_skill_set = args.skill_set
-
-    if target not in KNOWN_PROFILES:
-        store = AgentProfileStore()
-        try:
-            saved = store.resolve(target)
-        except AgentProfileError as e:
-            print(f"Error: {e}", file=sys.stderr)
-            sys.exit(1)
-        if saved is None:
-            print(
-                f"Error: Unknown profile or saved agent: {target}\n"
-                "Run 'synapse agents list' to see saved agent IDs/names.",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-        resolved_profile = saved.profile
-        resolved_name = resolved_name or saved.name
-        resolved_role = resolved_role or saved.role
-        resolved_skill_set = resolved_skill_set or saved.skill_set
-
-    if resolved_name:
-        registry = AgentRegistry()
-        if not registry.is_name_unique(resolved_name):
-            print(
-                f"Error: Name '{resolved_name}' is already taken by another agent.",
-                file=sys.stderr,
-            )
-            sys.exit(1)
-
-    try:
-        result = spawn_agent(
-            profile=resolved_profile,
-            port=args.port,
-            name=resolved_name,
-            role=resolved_role,
-            skill_set=resolved_skill_set,
-            terminal=args.terminal,
-            tool_args=tool_args or None,
-            worktree=getattr(args, "worktree", None),
-            auto_approve=not getattr(args, "no_auto_approve", False),
-            branch=getattr(args, "branch", None),
-        )
-        print(f"{result.agent_id} {result.port}")
-        if result.worktree_path:
-            print(f"  worktree: {result.worktree_path} ({result.worktree_branch})")
-
-        wait_timeout = getattr(args, "task_timeout", 30) if task_message else 3.0
-        info = wait_for_agent(result.agent_id, timeout=wait_timeout, poll_interval=0.5)
-        if info is None:
-            print(
-                f"Warning: {result.agent_id} not yet registered after spawn.\n"
-                f"  The agent may still be starting up.\n"
-                f"  Use the full agent ID for reliable targeting:\n"
-                f'    synapse send {result.agent_id} "<message>" --wait',
-                file=sys.stderr,
-            )
-            return
-
-        if task_message is not None:
-            cmd = _build_a2a_cmd(
-                "send",
-                task_message,
-                target=result.agent_id,
-                response_mode=getattr(args, "response_mode", None),
-                force=bool(result.worktree_path),
-            )
-            _run_a2a_command(cmd, exit_on_error=True)
-    except (FileNotFoundError, RuntimeError, ValueError) as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
 
 
 # ============================================================
